+++ src/main/java/com/techelevator/models/Candy.java
@@ -3,9 +3,11 @@ package com.techelevator.models;
 
 import com.techelevator.interfaces.TypeConstants;
 
+import java.math.BigDecimal;
+
 public class Candy extends Product implements TypeConstants {
 
-    public Candy(String name, double price) {
+    public Candy(String name, BigDecimal price) {
         super(name, price, TYPE_CANDY);
     }
