+++ src/main/java/com/techelevator/models/Chip.java
@@ -2,8 +2,10 @@ package com.techelevator.models;
 
 import com.techelevator.interfaces.TypeConstants;
 
+import java.math.BigDecimal;
+
 public class Chip extends Product implements TypeConstants {
-    public Chip(String name, double price) {
+    public Chip(String name, BigDecimal price) {
         super(name, price, TYPE_CHIP);
     }
